refactor: restructure game rooms
pom-lite will be the base for solo and 'simplified' multiplayer game
modes

all lobbies (including for pom-lite multiplayer will inherit from the
same base lobby)
