--- server/src/rooms/game/commands/types.ts
@@ -1,5 +1,0 @@
-import { GameEvent } from "@port-of-mars/server/rooms/game/events/types";
-
-export interface Command {
-  execute(): Array<GameEvent>;
-}
