+++ server/src/rooms/lobby/tournament/index.ts
@@ -3,7 +3,7 @@ import _ from "lodash";
 import { v4 as uuidv4 } from "uuid";
 import { Mutex } from "async-mutex";
 import { buildGameOpts } from "@port-of-mars/server/util";
-import { GameRoom } from "@port-of-mars/server/rooms/game";
+import { GameRoom } from "@port-of-mars/server/rooms/pom/game";
 import { settings } from "@port-of-mars/server/settings";
 import { getServices } from "@port-of-mars/server/services";
 import {
