+++ server/src/rooms/pom/game/commands/types.ts
@@ -1,0 +1,5 @@
+import { GameEvent } from "@port-of-mars/server/rooms/pom/game/events/types";
+
+export interface Command {
+  execute(): Array<GameEvent>;
+}
