+++ server/src/rooms/pom/game/index.ts
@@ -25,17 +25,17 @@ import {
   SetPlayerReadinessCmd,
   TimeInvestmentCmd,
   VoteForPhilanthropistCmd,
-} from "@port-of-mars/server/rooms/game/commands";
+} from "@port-of-mars/server/rooms/pom/game/commands";
 import { User } from "@port-of-mars/server/entity";
-import { Command } from "@port-of-mars/server/rooms/game/commands/types";
-import { TakenStateSnapshot } from "@port-of-mars/server/rooms/game/events";
-import { GameState, Player } from "@port-of-mars/server/rooms/game/state";
-import { Game, GameOpts, Metadata, Persister } from "@port-of-mars/server/rooms/game/types";
+import { Command } from "@port-of-mars/server/rooms/pom/game/commands/types";
+import { TakenStateSnapshot } from "@port-of-mars/server/rooms/pom/game/events";
+import { GameState, Player } from "@port-of-mars/server/rooms/pom/game/state";
+import { Game, GameOpts, Metadata, Persister } from "@port-of-mars/server/rooms/pom/game/types";
 import { getServices } from "@port-of-mars/server/services";
 import { settings } from "@port-of-mars/server/settings";
 import { Requests, Responses } from "@port-of-mars/shared/game";
 import { InspectData, Phase } from "@port-of-mars/shared/types";
-import { GameEvent } from "@port-of-mars/server/rooms/game/events/types";
+import { GameEvent } from "@port-of-mars/server/rooms/pom/game/events/types";
 import _ from "lodash";
 import { DBPersister } from "@port-of-mars/server/services/persistence";
 import { mockGameInitOpts } from "@port-of-mars/server/util";
