+++ server/src/rooms/pom/game/state/resource.ts
