+++ server/src/rooms/pom/game/types.ts
@@ -1,8 +1,8 @@
 import { Client, Room } from "colyseus";
-import { GameState, Player } from "@port-of-mars/server/rooms/game/state";
+import { GameState, Player } from "@port-of-mars/server/rooms/pom/game/state";
 import { Responses } from "@port-of-mars/shared/game/responses";
 import { GameType, MarsEventData, Role } from "@port-of-mars/shared/types";
-import * as ge from "@port-of-mars/server/rooms/game/events/types";
+import * as ge from "@port-of-mars/server/rooms/pom/game/events/types";
 import { GameEvent } from "@port-of-mars/server/entity/GameEvent";
 
 export type PlayerReadiness = { [role in Role]: boolean };
