+++ server/src/rooms/pomlite/multiplayer/index.ts
@@ -1,50 +1,62 @@
 import { Client, Delayed, Room } from "colyseus";
 import { Dispatcher } from "@colyseus/command";
 import * as http from "http";
-import { SoloGameState } from "@port-of-mars/server/rooms/sologame/state";
+import { TrioGameState, Player } from "@port-of-mars/server/rooms/pomlite/multiplayer/state";
 import { settings } from "@port-of-mars/server/settings";
 import { getServices } from "@port-of-mars/server/services";
-import { ApplyCardCmd, InitGameCmd, InvestCmd, PersistRoundCmd, SetNextRoundCmd } from "./commands";
+import {
+  ApplyCardCmd,
+  InitGameCmd,
+  SetPlayerCmd,
+  ProcessRoundCmd,
+  PlayerInvestCmd,
+} from "./commands";
 import { User } from "@port-of-mars/server/entity";
-import { EventContinue, Invest, SOLO_ROOM_NAME, SoloGameType } from "@port-of-mars/shared/sologame";
+import { EventContinue, Invest } from "@port-of-mars/shared/sologame";
+import { TrioGameOpts } from "@port-of-mars/server/rooms/pomlite/multiplayer/types";
 
 const logger = settings.logging.getLogger(__filename);
 
-export class SoloGameRoom extends Room<SoloGameState> {
+export class TrioGameRoom extends Room<TrioGameState> {
   public static get NAME() {
-    return SOLO_ROOM_NAME;
+    return "trio_game_room";
   }
+
   autoDispose = true;
-  maxClients = 1;
-  patchRate = 1000 / 5; // sends state to client 5 times per second
+  maxClients = 3;
+  patchRate = 1000 / 5;
 
   dispatcher = new Dispatcher(this);
   eventTimeout: Delayed | null = null;
 
-  get client() {
-    return this.clients[0];
-  }
-
-  onCreate(options: { type?: SoloGameType }) {
-    logger.trace("SoloGameRoom '%s' created", this.roomId);
-    this.setState(new SoloGameState());
+  async onCreate(options: TrioGameOpts) {
+    logger.trace("TrioGameRoom '%s' created", this.roomId);
+    this.setState(new TrioGameState(options));
     this.state.type = options.type || "freeplay";
     this.setPrivate(true);
     this.registerAllHandlers();
+    new InitGameCmd().setPayload({ users: options.users });
     this.clock.setInterval(() => {
       if (this.state.timeRemaining > 0) {
         this.state.timeRemaining -= 1;
       } else if (!this.state.isRoundTransitioning) {
-        this.dispatcher.dispatch(
-          new InvestCmd().setPayload({
-            systemHealthInvestment: 0,
-            clockRanOut: true,
-          })
-        );
+        this.state.players.forEach(player => {
+          player.pendingInvestment = null;
+          player.pointsEarned = null;
+          this.dispatcher.dispatch(
+            new PlayerInvestCmd().setPayload({
+              systemHealthInvestment: 0,
+              clockRanOut: true,
+              player: player,
+            })
+          );
+        });
+        this.dispatcher.dispatch(new ProcessRoundCmd());
       }
     }, 1000);
   }
 
+  // maybe add a safeguard to ensure client belongs to this room
   async onAuth(client: Client, options: any, request?: http.IncomingMessage) {
     try {
       const user = await getServices().account.findUserById((request as any).session.passport.user);
@@ -60,13 +72,17 @@ export class SoloGameRoom extends Room<SoloGameState> {
     return false;
   }
 
-  onJoin(client: Client, options: any, auth: User) {
-    logger.trace("Client %s joined SoloGameRoom %s", auth.username, this.roomId);
-    this.dispatcher.dispatch(new InitGameCmd().setPayload({ user: auth }));
+  getPlayer(client: Client): Player {
+    return this.state.getPlayer(client.auth.username);
   }
 
-  onDispose() {
-    logger.trace("Disposing of SoloGameRoom '%s'", this.roomId);
+  onJoin(client: Client, options: any, auth: User) {
+    logger.trace("Client %s joined TrioGameRoom %s", auth.username, this.roomId);
+    this.dispatcher.dispatch(new SetPlayerCmd().setPayload({ users: [auth] }));
+  }
+
+  async onDispose(): Promise<void> {
+    logger.trace("Disposing of TrioGameRoom '%s'", this.roomId);
     this.dispatcher.stop();
   }
 
@@ -75,8 +91,9 @@ export class SoloGameRoom extends Room<SoloGameState> {
       this.dispatcher.dispatch(new ApplyCardCmd().setPayload({ playerSkipped: true }));
     });
     this.onMessage("invest", (client, message: Invest) => {
+      this.state.getPlayer();
       this.dispatcher.dispatch(
-        new InvestCmd().setPayload({
+        new PlayerInvestCmd().setPayload({
           systemHealthInvestment: message.systemHealthInvestment,
         })
       );
