+++ server/src/rooms/pomlite/solo/state.ts
